Refactor: Update rating type from Double to Float
This commit updates the data type of `rating` from Double to Float across several files:

- `Stats.kt`: Changed `rating` property type.
- `UserRepositoryImpl.kt`: Updated default `rating` value.
- `UserRemoteDataSource.kt` and `UserRemoteDataSourceImpl.kt`:
    - Modified rating-related functions to use Float.
    - Added new functions for managing craftsman ratings, earnings, and jobs done.
- `CraftsmanHomeScreen.kt`: Updated preview data.
- `StatsContainer.kt`: Changed `rating` parameter type.
- `CraftsmanHomeViewModel.kt`: Adjusted mapping to `StatsUiState`.
- `CraftsmanHomeUiState.kt`:
    - Created `StatsUiState` with Float for rating.
    - Added `toStatsUiState` extension function.

+++ app/src/main/java/com/paris_2/san3a/domain/entity/Stats.kt
@@ -4,5 +4,5 @@ data class Stats(
     val userId: String,
     val jobsDone: Int,
     val earnings: Double,
-    val rating: Double,
+    val rating: Float,
 )
